+++ funciones_pila.h
@@ -5,7 +5,7 @@
 * \param uint8_t *memoria
 * \return Funcion tipo void pues modifica directamente el valor de la memoria
 */
-void PUSH(unsigned long registros[16], unsigned long *R, uint8_t *memoria);
+void PUSH(uint8_t registros[16], unsigned long *R, uint8_t *memoria);
 /**
 * \brief Funcion encargada de pasar datos memoria-registro
 * \param unsigned long registros[16]
@@ -13,4 +13,4 @@ void PUSH(unsigned long registros[16], unsigned long *R, uint8_t *memoria);
 * \param uint8_t *memoria
 * \return Funcion tipo void pues modifica directamente el valor de los registros
 */
-void POP(unsigned long registros[16], unsigned long *R, uint8_t *memoria);
+void POP(uint8_t registros[16], unsigned long *R, uint8_t *memoria);
